pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "varchiver"
version = "0.6.22"
description = "Advanced Archive Management Tool"
authors = [{ name = "instancer-kirik", email = "kirik@instance.select" }]
license = { text = "Proprietary" }
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "PyQt6>=6.7.1",
    "rarfile>=4.2.0",
    "pyqt6-sip>=13.8.0",
    "psutil>=6.1.0",
    "pyinstaller>=6.11.1",
    "requests>=2.32.3",
    "pillow>=11.1.0",
    "supabase>=2.15.1",
    "sqlalchemy>=2.0.41",
    "psycopg2-binary>=2.9.10",
    "PyYAML>=6.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2>=2.9.11",
]

[project.optional-dependencies]
dev = ["pyinstaller>=6.11.1", "pyarmor>=8.5.12"]

[project.scripts]
varchiver = "varchiver.main:main"
supamerge = "varchiver.supamerge.cli:cli_entry_point"

[tool.hatch.build.targets.wheel]
packages = ["varchiver"]
